Move WaitForEditors() to testUtils
